SearchLambda/requirements.txt
boto3
mysql-connector-python
weaviate-client
